TPT Billing Scenario update
Added dynamic routes for the review stage of a TPT bill run
3 issues:
Over abstraction, Incomplete returns and returns not processed
Options are to add default or custom quantities to the billable amount or you can remove items from the bill run.
The bill run review page should then reflect the changes. It does this through a JS script at the bottom of the page (bit clunky).
Scenarios can be reset by re-creating the TPT bill run