Updating charge information statuses
Amended approved and replaced to active. This is because they are both being used to calculate bills.

Also amended the grey to green so that we aren't using colour to determine meaning.

Active is either the most current charge information or previous charge information that is still being used for billing.
Invalid is used for charge information that is now overwritten and isn’t used for billing
Review, charge information that has been confirmed from draft and before it is made active