Sandbox Licence page update - alerts and errors pattern updates
Some work on bringing the licence page over to the sandbox, this is now based on the ID from the search results. Something to bear in mind when linking from other areas of the service.

System messages was also updated to include an error as well as an alert. I also amended the colours to reflect the latest accessiblity thinking in the design sytem. These should be updated when GDS updates the patterns around system notifications