Abs period added to station table
Following a three hats, we can now get the abs period consistently, so the table design has been simplified to include the abs period and to put flows and level in the same column.

This is because UR has shown that it is very unlikely that there will ever be a mix of flows and levels on the same monitoring station.