Contact details page formatting
Complicated table formatting for the contact details pages based on the different scenarios that we have.

two other versions of the contact details archived in case we want to simplify this.

Bug fix in the contact set up route, this was defaulting to customer0, it should now respect which customer you are adding a contact for.