Split bill runs, bills table creation
Created a more dynamic table of data for SRoC supplementary bill runs

Also tweaked the design of the linked component, after use it seemed counter intuitive to have the other charge scheme written on the bill run you were on. So i shifted this bit of info to the meta data.