Charge information workflow splitting the tabs
The tabbed view on the charge information workflow is causing issues with the page load. This is because all the date is loaded as one page on three tables.

This means the page load speed is low and the pagination isn't working correctly.

This feature update is to split the page onto separate pages using a contents component.

It also includes the billed up to date column on to set up.