Comparing changes
Designs looking at showing what has changed between versions.

This example is focusing on charge information.

It needs more exploration and looking at more scenarios. It might be better off to start with showing differences in abstraction data. Especially as this is created outside of MWAS and is used as a decision point for setting up new returns and charges.

Also included in this update is the use of the summary card for displaying charge information.