Show / hide voids
Void returns have become more common, they aren't necessary unless you are managing the day to day set up of a licence or making a historical correction to the data. Therefore the proposal is to hide them by default and have a feature to for the user to be able to show them if necessary.  I've used a button to implement this as it is an action and the table caption explains what has and hasn't happened. This does rely on the users knowing what a void return is.